requirements.txt
tensorflow==2.18.0
# tensorflow-metal==1.2.0 #! Only for Apple Silicon
Jinja2==3.1.5 
scikit-learn==1.6.1
uproot==5.5.1
vector==1.6.1
awkward==2.7.4
pandas==1.5.3
matplotlib==3.9.2 
seaborn==0.13.2
typed-argument-parser==1.10.1
hydra-core==1.3.2
pdoc3==0.11.5
atlas-ftag-tools==0.2.8 
atlasify==0.8.0
pytest==7.0.1
ray==2.42.1
awkward-pandas==2023.8.0
onnx==1.17.0
tf2onnx==1.16.1